Updating the visuals and background functions for canceling booking.
Now when you open the modal to book or cancel a booking, the modal submit button reflects what you want to do. The flash text is updated as well. I also updated how the home.html displays the room numbers. It is now dinamic and reflects what is stored in the database.